refactor(common): introduce rustfs-data-usage core crate
