+++ crates/common/Cargo.toml
@@ -34,11 +34,9 @@ uuid = { workspace = true }
 chrono = { workspace = true }
 metrics = { workspace = true }
 rustfs-madmin = { workspace = true }
-rustfs-filemeta = { workspace = true }
+rustfs-data-usage = { workspace = true }
 serde = { workspace = true }
-path-clean = { workspace = true }
 rmp-serde = { workspace = true }
-async-trait = { workspace = true }
 s3s = { workspace = true }
 tracing = { workspace = true }
