+++ crates/data-usage/src/lib.rs
@@ -12,4 +12,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-pub use rustfs_common::last_minute::*;
+pub mod data_usage;
+
+pub use data_usage::*;
